requirements.txt
requests
openai
Flask
python-dotenv